review folyamat 2.0 55%

+++ appointmentManagementAPI/src/main/java/com/Hbence/appointmentManagementAPI/repository/ReservationRepository.java
@@ -13,7 +13,7 @@ import java.util.List;
 public interface ReservationRepository extends JpaRepository<Reservations, Long> {
 
     @Procedure(name = "getReservationByUserId", procedureName = "getReservationByUserId")
-    List<Reservations> reservations(@Param("userIdIN") Integer userId);
+    List<Reservations> reservations(@Param("userIdIN") Long userId);
 
     @Procedure(name = "getReservationByDate", procedureName = "getReservationByDate")
     List<Reservations> getReservationByDate(@Param("dateIN") LocalDate wantedDate);
